Fix compile warnings and messages

+++ Client/mods/deathmatch/logic/CStaticFunctionDefinitions.cpp
@@ -1442,7 +1442,7 @@ bool CStaticFunctionDefinitions::SetElementModel(CClientEntity& Entity, unsigned
         {
             // Grab the model
             CClientPed& Ped = static_cast<CClientPed&>(Entity);
-            const unsigned short usCurrentModel = Ped.GetModel();
+            const unsigned short usCurrentModel = static_cast<ushort>(Ped.GetModel());
 
             if (usCurrentModel == usModel)
                 return false;
